Extract RB::CollisionElement into its own file.
