+++ drake/examples/kuka_iiwa_arm/iiwa_simulation.cc
@@ -40,7 +40,7 @@ std::shared_ptr<RigidBodySystem> CreateKukaIiwaSystem(void) {
   world.addVisualElement(
       DrakeShapes::VisualElement(geom, T_element_to_link, color));
   tree->addCollisionElement(
-      RigidBody::CollisionElement(geom, T_element_to_link, &world), world,
+      RigidBodyCollisionElement(geom, T_element_to_link, &world), world,
       "terrain");
   tree->updateStaticCollisionElements();
